restructure package organization for improved encapsulation
