+++ src/main/java/com/github/flo456123/BackBond/data/repository/EntryRepository.java
@@ -1,6 +1,6 @@
-package com.github.flo456123.BackBond.repository;
+package com.github.flo456123.BackBond.data.repository;
 
-import com.github.flo456123.BackBond.repository.model.Entry;
+import com.github.flo456123.BackBond.data.models.Entry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EntryRepository extends JpaRepository<Entry, Integer> {
